requirements.txt
pymupdf
sentence-transformers
faiss-cpu
gradio
python-dotenv
numpy
pandas
requests

